Move CscArguments and CscVsMSBuild tests to CscOnly category
Move the CscArguments and CscVsMSBuild tests from RunFileTests_Directives
to RunFileTests_CscOnlyAndApi as they are CscOnly tests, not Directive tests.
