+++ test/dotnet.Tests/CommandTests/Run/RunFileTests_Directives.cs
@@ -2,12 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections.Immutable;
-using System.Text.Json;
-using Basic.CompilerLog.Util;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.DotNet.Cli.Commands;
 using Microsoft.DotNet.Cli.Commands.Run;
-using Microsoft.DotNet.Cli.Utils;
 using Microsoft.DotNet.FileBasedPrograms;
 using Microsoft.DotNet.ProjectTools;
 
@@ -1132,512 +1129,6 @@ public sealed class RunFileTests_Directives(ITestOutputHelper log) : RunFileTest
             v2
             MySecret=MyValue (JsonConfigurationProvider for 'secrets.json' (Optional))
             """);
-    }
-
-    /// <summary>
-    /// Verifies that msbuild-based runs use CSC args equivalent to csc-only runs.
-    /// Can regenerate CSC arguments template in <see cref="CSharpCompilerCommand"/>.
-    /// </summary>
-    [Fact]
-    public void CscArguments()
-    {
-        var testInstance = TestAssetsManager.CreateTestDirectory(baseDirectory: OutOfTreeBaseDirectory);
-        const string programName = "TestProgram";
-        const string fileName = $"{programName}.cs";
-        string entryPointPath = Path.Join(testInstance.Path, fileName);
-        File.WriteAllText(entryPointPath, s_program);
-
-        // Remove artifacts from possible previous runs of this test.
-        var artifactsDir = VirtualProjectBuilder.GetArtifactsPath(entryPointPath);
-        if (Directory.Exists(artifactsDir)) Directory.Delete(artifactsDir, recursive: true);
-
-        // Build using MSBuild.
-        new DotnetCommand(Log, "run", fileName, "-bl", "--no-cache")
-            .WithWorkingDirectory(testInstance.Path)
-            .Execute()
-            .Should().Pass()
-            .And.HaveStdOut($"Hello from {programName}");
-
-        // Find the csc args used by the build.
-        var msbuildCall = FindCompilerCall(Path.Join(testInstance.Path, "msbuild.binlog"));
-        var msbuildCallArgs = msbuildCall.GetArguments();
-        var msbuildCallArgsString = ArgumentEscaper.EscapeAndConcatenateArgArrayForProcessStart(msbuildCallArgs);
-
-        // Generate argument template code.
-        string sdkPath = NormalizePath(SdkTestContext.Current.ToolsetUnderTest.SdkFolderUnderTest);
-        string dotNetRootPath = NormalizePath(SdkTestContext.Current.ToolsetUnderTest.DotNetRoot);
-        string nuGetCachePath = NormalizePath(SdkTestContext.Current.NuGetCachePath!);
-        string artifactsDirNormalized = NormalizePath(artifactsDir);
-        string objPath = $"{artifactsDirNormalized}/obj/debug";
-        string entryPointPathNormalized = NormalizePath(entryPointPath);
-        var msbuildArgsToVerify = new List<string>();
-        var nuGetPackageFilePaths = new List<string>();
-        bool referenceSpreadInserted = false;
-        bool analyzerSpreadInserted = false;
-        const string NetCoreAppRefPackPath = "packs/Microsoft.NETCore.App.Ref/";
-        var code = new StringBuilder();
-        code.AppendLine($$"""
-            // Licensed to the .NET Foundation under one or more agreements.
-            // The .NET Foundation licenses this file to you under the MIT license.
-
-            using System.Text.Json;
-
-            namespace Microsoft.DotNet.Cli.Commands.Run;
-
-            // Generated by test `{{nameof(RunFileTests)}}.{{nameof(CscArguments)}}`.
-            partial class CSharpCompilerCommand
-            {
-                private IEnumerable<string> GetCscArguments(
-                    string objDir,
-                    string binDir)
-                {
-                    return
-                    [
-            """);
-        foreach (var arg in msbuildCallArgs)
-        {
-            // This option needs to be passed on the command line, not in an RSP file.
-            if (arg is "/noconfig")
-            {
-                continue;
-            }
-
-            // We don't need to generate a ref assembly.
-            if (arg.StartsWith("/refout:", StringComparison.Ordinal))
-            {
-                continue;
-            }
-
-            // There should be no source link arguments.
-            if (arg.StartsWith("/sourcelink:", StringComparison.Ordinal))
-            {
-                Assert.Fail($"Unexpected source link argument: {arg}");
-            }
-
-            // PreferredUILang is normally not set by default but can be in builds, so ignore it.
-            if (arg.StartsWith("/preferreduilang:", StringComparison.Ordinal))
-            {
-                continue;
-            }
-
-            bool needsInterpolation = false;
-            bool fromNuGetPackage = false;
-
-            // Normalize slashes in paths.
-            string rewritten = NormalizePathArg(arg);
-
-            // Remove quotes.
-            rewritten = RemoveQuotes(rewritten);
-
-            string msbuildArgToVerify = rewritten;
-
-            // Use variable SDK path.
-            if (rewritten.Contains(sdkPath, StringComparison.OrdinalIgnoreCase))
-            {
-                rewritten = rewritten.Replace(sdkPath, "{SdkPath}", StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-            }
-
-            // Use variable .NET root path.
-            if (rewritten.Contains(dotNetRootPath, StringComparison.OrdinalIgnoreCase))
-            {
-                rewritten = rewritten.Replace(dotNetRootPath, "{DotNetRootPath}", StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-            }
-
-            // Use variable NuGet cache path.
-            if (rewritten.Contains(nuGetCachePath, StringComparison.OrdinalIgnoreCase))
-            {
-                rewritten = rewritten.Replace(nuGetCachePath, "{NuGetCachePath}", StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-                fromNuGetPackage = true;
-            }
-
-            // Use variable intermediate dir path.
-            if (rewritten.Contains(objPath, StringComparison.OrdinalIgnoreCase))
-            {
-                // We want to emit the resulting DLL directly into the bin folder.
-                bool isOut = arg.StartsWith("/out", StringComparison.Ordinal);
-                string replacement = isOut ? "{binDir}" : "{objDir}";
-
-                if (isOut)
-                {
-                    msbuildArgToVerify = msbuildArgToVerify.Replace("/obj/", "/bin/", StringComparison.OrdinalIgnoreCase);
-                }
-
-                rewritten = rewritten.Replace(objPath, replacement, StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-            }
-
-            // Use variable file path.
-            if (rewritten.Contains(entryPointPathNormalized, StringComparison.OrdinalIgnoreCase))
-            {
-                rewritten = rewritten.Replace(entryPointPathNormalized, "{" + nameof(CSharpCompilerCommand.EntryPointFileFullPath) + "}", StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-            }
-
-            // Use variable file name.
-            if (rewritten.Contains(fileName, StringComparison.OrdinalIgnoreCase))
-            {
-                rewritten = rewritten.Replace(fileName, "{FileName}", StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-            }
-
-            // Use variable program name.
-            if (rewritten.Contains(programName, StringComparison.OrdinalIgnoreCase))
-            {
-                rewritten = rewritten.Replace(programName, "{FileNameWithoutExtension}", StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-            }
-
-            // Use variable runtime version.
-            if (rewritten.Contains(CSharpCompilerCommand.RuntimeVersion, StringComparison.OrdinalIgnoreCase))
-            {
-                rewritten = rewritten.Replace(CSharpCompilerCommand.RuntimeVersion, "{" + nameof(CSharpCompilerCommand.RuntimeVersion) + "}", StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-            }
-
-            // Use variable target framework version.
-            if (rewritten.Contains(CSharpCompilerCommand.TargetFrameworkVersion, StringComparison.OrdinalIgnoreCase))
-            {
-                rewritten = rewritten.Replace(CSharpCompilerCommand.TargetFrameworkVersion, "{" + nameof(CSharpCompilerCommand.TargetFrameworkVersion) + "}", StringComparison.OrdinalIgnoreCase);
-                needsInterpolation = true;
-            }
-
-            // Ignore `/analyzerconfig` which is not variable (so it comes from the machine or sdk repo).
-            if (!needsInterpolation && arg.StartsWith("/analyzerconfig", StringComparison.Ordinal))
-            {
-                continue;
-            }
-
-            // Use GetFrameworkReferenceArguments() for framework references instead of hard-coding them.
-            if (arg.StartsWith("/reference:", StringComparison.Ordinal))
-            {
-                if (!referenceSpreadInserted)
-                {
-                    code.AppendLine("""
-                                    .. GetFrameworkReferenceArguments(),
-                        """);
-                    referenceSpreadInserted = true;
-                }
-
-                msbuildArgsToVerify.Add(msbuildArgToVerify);
-                continue;
-            }
-
-            // Use GetFrameworkAnalyzerArguments() for targeting-pack analyzers instead of hard-coding them.
-            if (arg.StartsWith("/analyzer:", StringComparison.Ordinal)
-                && rewritten.Contains(NetCoreAppRefPackPath, StringComparison.OrdinalIgnoreCase))
-            {
-                if (!analyzerSpreadInserted)
-                {
-                    code.AppendLine("""
-                                    .. GetFrameworkAnalyzerArguments(),
-                        """);
-                    analyzerSpreadInserted = true;
-                }
-
-                msbuildArgsToVerify.Add(msbuildArgToVerify);
-                continue;
-            }
-
-            string prefix = needsInterpolation ? "$" : string.Empty;
-
-            code.AppendLine($"""
-                            {prefix}"{rewritten}",
-                """);
-
-            msbuildArgsToVerify.Add(msbuildArgToVerify);
-
-            if (fromNuGetPackage)
-            {
-                nuGetPackageFilePaths.Add(CSharpCompilerCommand.IsPathOption(rewritten, out int colonIndex)
-                    ? rewritten.Substring(colonIndex + 1)
-                    : rewritten);
-            }
-        }
-        code.AppendLine("""
-                    ];
-                }
-
-                /// <summary>
-                /// Files that come from referenced NuGet packages (e.g., analyzers for NativeAOT) need to be checked specially (if they don't exist, MSBuild needs to run).
-                /// </summary>
-                public static IEnumerable<string> GetPathsOfCscInputsFromNuGetCache()
-                {
-                    return
-                    [
-            """);
-        foreach (var nuGetPackageFilePath in nuGetPackageFilePaths)
-        {
-            code.AppendLine($"""
-                            $"{nuGetPackageFilePath}",
-                """);
-        }
-        code.AppendLine("""
-                    ];
-                }
-            """);
-
-        // Generate file content templates.
-        var baseDirectory = TestPathUtility.ResolveTempPrefixLink(Path.GetDirectoryName(entryPointPath)!);
-        var replacements = new List<(string, string)>
-        {
-            (TestPathUtility.ResolveTempPrefixLink(entryPointPath), nameof(CSharpCompilerCommand.EntryPointFileFullPath)),
-            (baseDirectory + Path.DirectorySeparatorChar, nameof(CSharpCompilerCommand.BaseDirectoryWithTrailingSeparator)),
-            (baseDirectory, nameof(CSharpCompilerCommand.BaseDirectory)),
-            (programName, nameof(CSharpCompilerCommand.FileNameWithoutExtension)),
-            (CSharpCompilerCommand.TargetFrameworkVersion, nameof(CSharpCompilerCommand.TargetFrameworkVersion)),
-            (CSharpCompilerCommand.TargetFramework, nameof(CSharpCompilerCommand.TargetFramework)),
-            (CSharpCompilerCommand.DefaultRuntimeVersion, nameof(CSharpCompilerCommand.DefaultRuntimeVersion)),
-        };
-        var emittedFiles = Directory.EnumerateFiles(artifactsDir, "*", SearchOption.AllDirectories).Order();
-        foreach (var emittedFile in emittedFiles)
-        {
-            var emittedFileName = Path.GetFileName(emittedFile);
-            var generatedMethodName = GetGeneratedMethodName(emittedFileName);
-            if (generatedMethodName is null)
-            {
-                Log.WriteLine($"Skipping unrecognized file '{emittedFile}'.");
-                continue;
-            }
-
-            var emittedFileContent = File.ReadAllText(emittedFile);
-
-            string interpolatedString = emittedFileContent;
-            string interpolationPrefix;
-
-            if (emittedFileName.EndsWith(".json", StringComparison.Ordinal))
-            {
-                interpolationPrefix = "$$";
-                foreach (var (key, value) in replacements)
-                {
-                    interpolatedString = interpolatedString.Replace(JsonSerializer.Serialize(key), "{{JsonSerializer.Serialize(" + value + ", CSharpCompilerCommandJsonSerializerContext.Default.String)}}");
-                }
-            }
-            else
-            {
-                interpolationPrefix = "$";
-                foreach (var (key, value) in replacements)
-                {
-                    interpolatedString = interpolatedString.Replace(key, "{" + value + "}");
-                }
-            }
-
-            if (interpolatedString == emittedFileContent)
-            {
-                interpolationPrefix = "";
-            }
-
-            code.AppendLine($$""""
-
-                    private string Get{{generatedMethodName}}Content()
-                    {
-                        return {{interpolationPrefix}}"""
-                {{interpolatedString}}
-                """;
-                    }
-                """");
-        }
-
-        code.AppendLine("""
-            }
-            """);
-
-        // Save the code.
-        var codeFolder = new DirectoryInfo(Path.Join(
-            SdkTestContext.Current.ToolsetUnderTest.RepoRoot,
-            "src", "Cli", "dotnet", "Commands", "Run"));
-        var nonGeneratedFile = codeFolder.File("CSharpCompilerCommand.cs");
-        if (!nonGeneratedFile.Exists)
-        {
-            Log.WriteLine($"Skipping code generation because file does not exist: {nonGeneratedFile.FullName}");
-        }
-        else
-        {
-            var codeFilePath = codeFolder.File("CSharpCompilerCommand.Generated.cs");
-            var existingText = codeFilePath.Exists ? File.ReadAllText(codeFilePath.FullName) : string.Empty;
-            var newText = code.ToString();
-            if (existingText != newText)
-            {
-                Log.WriteLine($"{codeFilePath.FullName} needs to be updated:");
-                Log.WriteLine(newText);
-                if (Env.GetEnvironmentVariableAsBool("CI"))
-                {
-                    throw new InvalidOperationException($"Not updating file in CI: {codeFilePath.FullName}");
-                }
-                else
-                {
-                    File.WriteAllText(codeFilePath.FullName, newText);
-                    throw new InvalidOperationException($"File outdated, commit the changes: {codeFilePath.FullName}");
-                }
-            }
-        }
-
-        // Build using CSC.
-        Directory.Delete(artifactsDir, recursive: true);
-        new DotnetCommand(Log, "run", fileName, "-bl")
-            .WithWorkingDirectory(testInstance.Path)
-            .Execute()
-            .Should().Pass()
-            .And.HaveStdOut($"""
-                {CliCommandStrings.NoBinaryLogBecauseRunningJustCsc}
-                Hello from {programName}
-                """);
-
-        // Read args from csc.rsp file.
-        var rspFilePath = Path.Join(artifactsDir, "csc.rsp");
-        var cscOnlyCallArgs = File.ReadAllLines(rspFilePath);
-        var cscOnlyCallArgsString = string.Join(' ', cscOnlyCallArgs);
-
-        // Check that csc args between MSBuild run and CSC-only run are equivalent.
-        var normalizedCscOnlyArgs = cscOnlyCallArgs
-            .Select(static a => NormalizePathArg(RemoveQuotes(a)))
-            .ToList();
-        Log.WriteLine("CSC-only args:");
-        Log.WriteLine(string.Join(Environment.NewLine, normalizedCscOnlyArgs));
-        Log.WriteLine("MSBuild args:");
-        Log.WriteLine(string.Join(Environment.NewLine, msbuildArgsToVerify));
-
-        // References and targeting-pack analyzers may be in a different order (FrameworkList.xml vs. MSBuild),
-        // so compare them as sets. All other args must be in the same order.
-        var cscOnlyRefArgs = normalizedCscOnlyArgs.Where(static a => a.StartsWith("/reference:", StringComparison.Ordinal)).ToList();
-        var cscOnlyAnalyzerArgs = normalizedCscOnlyArgs.Where(a => a.StartsWith("/analyzer:", StringComparison.Ordinal) && a.Contains(NetCoreAppRefPackPath, StringComparison.OrdinalIgnoreCase)).ToList();
-        var cscOnlyOtherArgs = normalizedCscOnlyArgs.Where(a => !a.StartsWith("/reference:", StringComparison.Ordinal) && !(a.StartsWith("/analyzer:", StringComparison.Ordinal) && a.Contains(NetCoreAppRefPackPath, StringComparison.OrdinalIgnoreCase))).ToList();
-        var msbuildRefArgs = msbuildArgsToVerify.Where(static a => a.StartsWith("/reference:", StringComparison.Ordinal)).ToList();
-        var msbuildAnalyzerArgs = msbuildArgsToVerify.Where(a => a.StartsWith("/analyzer:", StringComparison.Ordinal) && a.Contains(NetCoreAppRefPackPath, StringComparison.OrdinalIgnoreCase)).ToList();
-        var msbuildOtherArgs = msbuildArgsToVerify.Where(a => !a.StartsWith("/reference:", StringComparison.Ordinal) && !(a.StartsWith("/analyzer:", StringComparison.Ordinal) && a.Contains(NetCoreAppRefPackPath, StringComparison.OrdinalIgnoreCase))).ToList();
-        cscOnlyRefArgs.Should().NotBeEmpty(
-            "framework references should be resolved from FrameworkList.xml");
-        cscOnlyRefArgs.Should().BeEquivalentTo(msbuildRefArgs,
-            "the generated file might be outdated, run this test locally to regenerate it");
-        cscOnlyAnalyzerArgs.Should().NotBeEmpty(
-            "framework analyzers should be resolved from FrameworkList.xml");
-        cscOnlyAnalyzerArgs.Should().BeEquivalentTo(msbuildAnalyzerArgs,
-            "the generated file might be outdated, run this test locally to regenerate it");
-        cscOnlyOtherArgs.Should().Equal(msbuildOtherArgs,
-            "the generated file might be outdated, run this test locally to regenerate it");
-
-        static CompilerCall FindCompilerCall(string binaryLogPath)
-        {
-            using var reader = BinaryLogReader.Create(binaryLogPath);
-            return reader.ReadAllCompilerCalls().Should().ContainSingle().Subject;
-        }
-
-        static string NormalizePathArg(string arg)
-        {
-            return CSharpCompilerCommand.IsPathOption(arg, out int colonIndex)
-                ? string.Concat(arg.AsSpan(0, colonIndex + 1), NormalizePath(arg.Substring(colonIndex + 1)))
-                : NormalizePath(arg);
-        }
-
-        static string NormalizePath(string path)
-        {
-            return PathUtility.GetPathWithForwardSlashes(TestPathUtility.ResolveTempPrefixLink(path));
-        }
-
-        static string RemoveQuotes(string arg)
-        {
-            return arg.Replace("\"", string.Empty);
-        }
-
-        static string? GetGeneratedMethodName(string assetFileName)
-        {
-            return assetFileName switch
-            {
-                $".NETCoreApp,Version=v{ToolsetInfo.CurrentTargetFrameworkVersion}.AssemblyAttributes.cs" => "AssemblyAttributes",
-                $"{fileName}.GlobalUsings.g.cs" => "GlobalUsings",
-                $"{fileName}.AssemblyInfo.cs" => "AssemblyInfo",
-                $"{fileName}.GeneratedMSBuildEditorConfig.editorconfig" => "GeneratedMSBuildEditorConfig",
-                $"{programName}{FileNameSuffixes.RuntimeConfigJson}" => "RuntimeConfig",
-                _ => null,
-            };
-        }
-    }
-
-    /// <summary>
-    /// Verifies that csc-only runs emit auxiliary files equivalent to msbuild-based runs.
-    /// </summary>
-    [Theory]
-    [InlineData("Program.cs")]
-    [InlineData("test.cs")]
-    [InlineData("noext")]
-    public void CscVsMSBuild(string fileName)
-    {
-        var testInstance = TestAssetsManager.CreateTestDirectory(baseDirectory: OutOfTreeBaseDirectory);
-        string entryPointPath = Path.Join(testInstance.Path, fileName);
-        File.WriteAllText(entryPointPath, $"""
-            #!/test
-            {s_program}
-            """);
-
-        string programName = Path.GetFileNameWithoutExtension(fileName);
-
-        // Remove artifacts from possible previous runs of this test.
-        var artifactsDir = VirtualProjectBuilder.GetArtifactsPath(entryPointPath);
-        if (Directory.Exists(artifactsDir)) Directory.Delete(artifactsDir, recursive: true);
-        var artifactsBackupDir = Path.ChangeExtension(artifactsDir, ".bak");
-        if (Directory.Exists(artifactsBackupDir)) Directory.Delete(artifactsBackupDir, recursive: true);
-
-        // Build using CSC.
-        new DotnetCommand(Log, "run", fileName, "-bl")
-            .WithWorkingDirectory(testInstance.Path)
-            .Execute()
-            .Should().Pass()
-            .And.HaveStdOut($"""
-                {CliCommandStrings.NoBinaryLogBecauseRunningJustCsc}
-                Hello from {programName}
-                """);
-
-        // Backup the artifacts directory.
-        Directory.Move(artifactsDir, artifactsBackupDir);
-
-        // Build using MSBuild.
-        new DotnetCommand(Log, "run", fileName, "-bl", "--no-cache")
-            .WithWorkingDirectory(testInstance.Path)
-            .Execute()
-            .Should().Pass()
-            .And.HaveStdOut($"Hello from {programName}");
-
-        // Check that files generated by MSBuild and CSC-only runs are equivalent.
-        var cscOnlyFiles = Directory.EnumerateFiles(artifactsBackupDir, "*", SearchOption.AllDirectories)
-            .Where(f =>
-                Path.GetDirectoryName(f) != artifactsBackupDir && // exclude top-level marker files
-                Path.GetFileName(f) != programName && // binary on unix
-                Path.GetExtension(f) is not (".dll" or ".exe" or ".pdb")); // other binaries
-        bool hasErrors = false;
-        foreach (var cscOnlyFile in cscOnlyFiles)
-        {
-            var relativePath = Path.GetRelativePath(relativeTo: artifactsBackupDir, path: cscOnlyFile);
-            var msbuildFile = Path.Join(artifactsDir, relativePath);
-
-            if (!File.Exists(msbuildFile))
-            {
-                throw new InvalidOperationException($"File exists in CSC-only run but not in MSBuild run: {cscOnlyFile}");
-            }
-
-            var cscOnlyFileText = File.ReadAllText(cscOnlyFile);
-            var msbuildFileText = File.ReadAllText(msbuildFile);
-            if (cscOnlyFileText.ReplaceLineEndings() != msbuildFileText.ReplaceLineEndings())
-            {
-                Log.WriteLine($"File differs between MSBuild and CSC-only runs (if this is expected, run test '{nameof(CscArguments)}' locally to re-generate the template): {cscOnlyFile}");
-                const int limit = 3_000;
-                if (cscOnlyFileText.Length < limit && msbuildFileText.Length < limit)
-                {
-                    Log.WriteLine("MSBuild file content:");
-                    Log.WriteLine(msbuildFileText);
-                    Log.WriteLine("CSC-only file content:");
-                    Log.WriteLine(cscOnlyFileText);
-                }
-                else
-                {
-                    Log.WriteLine($"MSBuild file size: {msbuildFileText.Length} chars");
-                    Log.WriteLine($"CSC-only file size: {cscOnlyFileText.Length} chars");
-                }
-                hasErrors = true;
-            }
-        }
-        hasErrors.Should().BeFalse("some file contents do not match, see the test output for details");
     }
 
 }
